crash: add static_cast
The new version of Clang warns when a value of `enum class` type is
passed to a `printf`-like function without an explicit cast. This adds
the cast that Clang suggests.

BUG=b:305714217
TEST=emerge-hana

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5008817
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Auto-Submit: George Burgess <gbiv@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>

+++ crash-reporter/bluetooth_devcd_parser_util.cc
@@ -732,8 +732,9 @@ bool ReportDefaultPC(base::File& file, std::string* pc) {
 }
 
 bool ReportParseError(ParseErrorReason error_code, base::File& file) {
-  std::string line = CreateDumpEntry("Parse Failure Reason",
-                                     base::StringPrintf("%d", error_code));
+  std::string line =
+      CreateDumpEntry("Parse Failure Reason",
+                      base::StringPrintf("%d", static_cast<int>(error_code)));
   return file.WriteAtCurrentPosAndCheck(base::as_bytes(base::make_span(line)));
 }
